+++ packages/main/src/YearPicker.js
@@ -1,3 +1,4 @@
+import Integer from "@ui5/webcomponents-base/dist/types/Integer.js";
 import DateFormat from "@ui5/webcomponents-localization/dist/DateFormat.js";
 import {
 	isEnter,
@@ -16,7 +17,7 @@ import {
 import getLocale from "@ui5/webcomponents-base/dist/locale/getLocale.js";
 import CalendarDate from "@ui5/webcomponents-localization/dist/dates/CalendarDate.js";
 import { getMaxCalendarDate } from "@ui5/webcomponents-localization/dist/dates/ExtremeDates.js";
-import PickerBase from "./PickerBase.js";
+import CalendarPickerBase from "./CalendarPickerBase.js";
 import YearPickerTemplate from "./generated/templates/YearPickerTemplate.lit.js";
 import styles from "./generated/themes/YearPicker.css.js";
 
@@ -26,6 +27,26 @@ import styles from "./generated/themes/YearPicker.css.js";
 const metadata = {
 	tag: "ui5-yearpicker",
 	properties: /** @lends  sap.ui.webcomponents.main.YearPicker.prototype */ {
+		/**
+		 * The timestamp of the currently focused date
+		 * @type {Integer}
+		 * @public
+		 */
+		timestamp: {
+			type: Integer,
+		},
+
+		/**
+		 * An array of UTC timestamps representing the selected dates.
+		 * @type {Array}
+		 * @public
+		 */
+		selectedDates: {
+			type: Integer,
+			multiple: true,
+			compareValues: true,
+		},
+
 		_years: {
 			type: Object,
 			multiple: true,
@@ -44,7 +65,7 @@ const metadata = {
 		 */
 		change: {},
 		/**
-		 * Fired when the user navigates with the keyboard.
+		 * Fired when the timestamp changes - the user navigates with the keyboard or clicks with the mouse.
 		 * @since 1.0.0-rc.9
 		 * @public
 		 * @event
@@ -64,11 +85,11 @@ const ROW_SIZE = 4; // Years per row (5 rows of 4 years each)
  * @constructor
  * @author SAP SE
  * @alias sap.ui.webcomponents.main.YearPicker
- * @extends sap.ui.webcomponents.main.PickerBase
+ * @extends sap.ui.webcomponents.main.CalendarPickerBase
  * @tagname ui5-yearpicker
  * @public
  */
-class YearPicker extends PickerBase {
+class YearPicker extends CalendarPickerBase {
 	static get metadata() {
 		return metadata;
 	}
